FIXED compilation warning

+++ Classes/GON_NSString+Utils.m
@@ -26,7 +26,7 @@
 {
     NSMutableString *output = [NSMutableString string];
     const unsigned char* source = (const unsigned char *)[self UTF8String];
-    int sourceLen = strlen((const char *)source);
+    int sourceLen = (int)strlen((const char *)source);
     for (int i = 0; i < sourceLen; ++i) {
         const unsigned char thisChar = source[i];
         if (thisChar == ' '){
